[Codecov] Move Codecov logic in its own file
